Alt/requirements.txt
bottle==0.12.25
psycopg2-binary==2.9.9
requests==2.31.0
SQLAlchemy==2.0.23
jinja2==3.1.2
